requirements.txt
discord.py==1.4.1
pyquery==1.4.1